% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_eto_FAO56.R
\name{daily_eto_FAO56}
\alias{daily_eto_FAO56}
\title{Eto calculation based on FAO-56 methodology, with data from automatic weather stations (AWS) downloaded and processed in function *daily_download_AWS_INMET*
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
daily_eto_FAO56(lat, tmin, tmax, tmean, Rs, u2, P, RH_max, RH_min, z, date)
}
\arguments{
\item{lat}{A numeric value of the Latitude of the AWS in decimal degrees}

\item{tmin}{A dataframe with Minimum daily air temperature (°C)}

\item{tmax}{A dataframe with Maximum daily air temperature (°C)}

\item{tmean}{A dataframe with Mean daily air temperature (°C)}

\item{Rs}{A dataframe with mean daily solar radiation (MJ m-2 day-1)}

\item{u2}{A dataframe with Wind speed at meters high (m s-2)}

\item{P}{A dataframe with atmospheric Pressure (mmHg)}

\item{RH_max}{A dataframe with Maximum relative humidity}

\item{RH_min}{A dataframe with Minimum relative humidity}

\item{z}{A numeric value of the altitude of AWS}

\item{date}{A dataframe with the date information}
}
\value{
Returns a dataframe with the AWS data requested
}
\description{
This function will calculate the reference evapotranspiration (ETo) based on FAO-56 (Allen et al., 1998) with the the automatic weather stations (AWS) downloaded and processed in function *daily_download_AWS_INMET*.
}
\examples{
\dontrun{
eto<-daily_eto_FAO56(lat, tmin, tmax, tmean, Rs, u2, P, RH_max, RH_min, z, date)
}
}
\author{
Roberto Filgueiras
}
