% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_download_AWS_INMET.R
\name{download_AWS_INMET_daily}
\alias{download_AWS_INMET_daily}
\title{Download the hourly automatic weather station data of INMET and aggregate it in a daily time steps
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
download_AWS_INMET_daily(station, start_date, end_date)
}
\arguments{
\item{station}{The station identification (ID - WMO code) for download. To see the station ID, please see the function *see_stations_info*.}

\item{start_date}{Date that start the investigation, should be in the following format (1958-01-01 /Year-Month-Day)}

\item{end_date}{Date that end the investigation, should be in the following format (2017-12-31 /Year-Month-Day)}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
This function will download the hourly AWS data of INMET and it will aggregate the data in a daily time step, based on the period of time selected (start_date and end_date).The function only works for downloading data from the same year.
}
\examples{
\dontrun{
df<-download_AWS_INMET_daily(station = "A001", start_date = "2001-01-01", end_date = "2001-12-31")
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
