% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_air_humidity.R
\name{ea_dew_calculation}
\alias{ea_dew_calculation}
\title{Actual vapour pressure (ea) derived from dewpoint temperature}
\usage{
ea_dew_calculation(tdew)
}
\arguments{
\item{tdew}{A dataframe with dewpoint temperature (Celsius).}
}
\value{
Returns a data.frame object with the ea from dewpoint data.
}
\description{
Actual vapour pressure (ea) derived from dewpoint temperature
}
\examples{
\dontrun{
ea <-ea_dew_calculation(tdew)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha.
}
