% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eto_hs.R
\name{eto_hs}
\alias{eto_hs}
\title{Hargreaves - Samani ETo}
\usage{
eto_hs(tmin, tmean, tmax, ra)
}
\arguments{
\item{tmin}{A numeric object with Maximum daily air temperature (Celsius)}

\item{tmean}{A numeric object with Mean daily air temperature (Celsius)}

\item{tmax}{A numeric object with Maximum daily air temperature (Celsius)}

\item{ra}{A numeric object of extraterrestrial radiation (MJ m-2 day-1)}
}
\value{
Returns a numeric object with the ETo HS data
}
\description{
Hargreaves - Samani ETo
}
\examples{
\dontrun{
eto_hs <- eto_hs(tmin, tmean, tmax, ra)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
