% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_thorntwaite.R
\name{etp_thorntwaite}
\alias{etp_thorntwaite}
\title{Thorntwaite - Potential evapotranspiration}
\usage{
etp_thorntwaite(tmean)
}
\arguments{
\item{tmean}{A dataframe with Mean monthly air temperature (°C)}
}
\value{
Returns a data.frame object with the Thorntwaite ETp data
}
\description{
Thorntwaite - Potential evapotranspiration
}
\examples{
\dontrun{
etp <- etp_thorntwaite(tmean)
}
}
\author{
Roberto Filgueiras.
}
