% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_parameters.R
\name{psy_const}
\alias{psy_const}
\title{Psychrometric constant}
\usage{
psy_const(Patm)
}
\arguments{
\item{Patm}{Atmospheric pressure (kPa)}
}
\value{
A data.frame object with the psychrometric constant calculated.
}
\description{
Psychrometric constant (kPa/Celsius) is calculated in this function.
}
\examples{
\dontrun{
psy_df <- psy_const(Patm)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
