% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_parameters.R
\name{rnl_calculation}
\alias{rnl_calculation}
\title{Net longwave radiation (rnl)}
\usage{
rnl_calculation(tmin, tmax, ea, rs, rso)
}
\arguments{
\item{tmin}{A dataframe with Minimum daily air temperature (°C)}

\item{tmax}{A dataframe with Maximum daily air temperature (°C)}

\item{ea}{A dataframe with the actual vapour pressure (KPa).}

\item{rs}{A dataframe with the incomimg solar radiation (MJ m-2 day-1).}

\item{rso}{A dataframe with the clear-sky radiation (MJ m-2 day-1)}
}
\value{
A data.frame object with the net longwave radiation.
}
\description{
Net outgoing longwave radiation is calculate with this function
}
\examples{
\dontrun{
rnl_df <- rnl_calculation(tmin, tmax, ea, rs, rso)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
