% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_parameters.R
\name{rns_calculation}
\alias{rns_calculation}
\title{Net solar or net shortwave radiation (rns)
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
rns_calculation(albedo, rs)
}
\arguments{
\item{albedo}{Albedo or canopy reflectance coefficient. The 0.23 is the value used for hypothetical grass reference crop (dimensionless) .}

\item{rs}{The incomimg solar radiation (MJ m-2 day-1).}
}
\description{
The rns results form the balance between incoming and reflected solar radiation (MJ m-2 day-1).
}
\examples{
\dontrun{
ra <- rns_calculation(albedo, rs)
}
}
