% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_parameters.R
\name{rso_calculation_2}
\alias{rso_calculation_2}
\title{Clear-sky solar radiation when calibrated values are not available
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
rso_calculation_2(z, ra)
}
\arguments{
\item{z}{Station elevation above sea level (m)}

\item{ra}{Extraterrestrial radiation for daily periods (ra).}
}
\value{
A data.frame object with the clear-sky solar radiation
}
\description{
Clear-sky solar radiation is calculated in this function for near sea level or when calibrated values for as and bs are available.
}
\examples{
\dontrun{
rso_df <- rso_calculation_2(z, ra)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
