% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_stations_info.R
\name{see_stations_info}
\alias{see_stations_info}
\title{Localization of the automatic weather station of INMET}
\usage{
see_stations_info()
}
\value{
A data.frame with informations of OMM code, latitude, longitude and altitude of all AWS stations available in INMET.
}
\description{
Function to see the localization of the automatic weather station of INMET.
}
\examples{
\dontrun{
see_stations_info()
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
