% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAWSstations.R
\name{selectAWSstations}
\alias{selectAWSstations}
\title{Select Automatic Weather Stations}
\usage{
selectAWSstations(aoi, as_sf = FALSE)
}
\arguments{
\item{aoi}{sf object. Area of interest.}

\item{as_sf}{logical. Inmet stations inventory result should be export as sf object. Default = FALSE.}
}
\value{
a vector or sf object with station located at area of interest (aoi)
}
\description{
Select automatic weather stations of INMET based on sf object.
}
\examples{
\dontrun{

es = geobr::read_municipality(code_muni = "ES")

esStations = selectStations(aoi = es, as_sf = F); es


}
}
