% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_speed_variables.R
\name{u2_calculation}
\alias{u2_calculation}
\title{Wind speed at 2 meters high}
\usage{
u2_calculation(uz, z)
}
\arguments{
\item{uz}{measured wind speed at z meters above ground surface}

\item{z}{height of measurement above ground surface.}
}
\value{
A data.frame with the wind speed at 2 meters high calculated.
}
\description{
Wind speed at two meters high can be calculated with this function.
}
\examples{
\dontrun{
u2_df <- u2_calculation(uz, z)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
