% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_air_humidity_2.R
\name{ea_rh_calculation}
\alias{ea_rh_calculation}
\title{Actual vapour pressure (ea) derived from relative humidity data}
\usage{
ea_rh_calculation(tmin, tmax, rh_min, rh_mean, rh_max)
}
\arguments{
\item{tmin}{A dataframe with minimum daily air temperature (°C)}

\item{tmax}{A dataframe with maximum daily air temperature (°C)}

\item{rh_min}{A dataframe with minimum daily relative air humidity}

\item{rh_mean}{A dataframe with mean daily relative air humidity}

\item{rh_max}{A dataframe with maximum daily relative air humidity}
}
\value{
Returns a data.frame object with the with ea from relative humidity data.
}
\description{
Actual vapour pressure (ea) derived from relative humidity data
}
\examples{
\dontrun{
ea <- ea_rh_calculation(tmin, tmax, rh_min, rh_mean, rh_max)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
