% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_air_humidity.R
\name{es_calculation}
\alias{es_calculation}
\title{Mean saturation vapour pressure (es)}
\usage{
es_calculation(tmin, tmax)
}
\arguments{
\item{tmin}{A dataframe with Minimum daily air temperature (°C)}

\item{tmax}{A dataframe with Maximum daily air temperature (°C)}
}
\value{
Returns a data.frame object with the with es data
}
\description{
Mean saturation vapour pressure (es)
}
\examples{
\dontrun{
es <-es_calculation(tmin, tmax)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
