% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eto_hs.R
\name{eto_hs}
\alias{eto_hs}
\title{Hargreaves - Samani ETo}
\usage{
eto_hs(tmin, tmean, tmax, ra)
}
\arguments{
\item{tmin}{A dataframe with Maximum daily air temperature (°C)}

\item{tmean}{A dataframe with Minimum daily air temperature (°C)}

\item{tmax}{A dataframe with Maximum daily air temperature (°C)}

\item{ra}{A dataframe of extraterrestrial radiation}
}
\value{
Returns a data.frame object with the with ETo HS data
}
\description{
Hargreaves - Samani ETo
}
\examples{
\dontrun{
eto_hs <-eto_hs(tmin, tmean, tmax, ra)
}
}
