% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_parameters.R
\name{P}
\alias{P}
\title{Atmospheric pressure (P)}
\usage{
P(z)
}
\arguments{
\item{z}{Elevation above sea level (m)}
}
\value{
Returns a data.frame object with the atmospheric pressure calculated.
}
\description{
Atmospheric pressure (P)
}
\examples{
\dontrun{
P <- p(z)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
