% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_parameters.R
\name{psy_const}
\alias{psy_const}
\title{Psychrometric constant}
\usage{
psy_const(P)
}
\arguments{
\item{P}{Atmospheric pressure (kPa)}
}
\description{
Psychrometric constant (kPa/°C) is calculated in this function.
}
\examples{
\dontrun{
psy_df <- psy_const(P)
}
}
