% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_parameters.R
\name{ra_calculation}
\alias{ra_calculation}
\title{Extraterrestrial radiation for daily periods (ra)
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
ra_calculation(latitude, date)
}
\arguments{
\item{latitude}{A dataframe with latitude in decimal degrees that you want to calculate the ra.}

\item{date}{A dataframe with the dates that you want to calculate the ra.}
}
\description{
ra is expressed in MJ m-2 day-1
}
\examples{
\dontrun{
ra <- ra_calculation(latitude, date)
}
}
