% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_air_humidity_2.R
\name{rh_calculation}
\alias{rh_calculation}
\title{Relative humidity (rh) calculation}
\usage{
rh_calculation(tmin, t)
}
\arguments{
\item{tmin}{A dataframe with minimum daily air temperature (°C)}

\item{t}{A dataframe with air temperature (°C) of the moment that you want to calculate the relative humidity.}
}
\description{
Psychrometric constant (kPa/°C) is calculated in this function.
}
\examples{
\dontrun{
rh <- rh_calculation(tmin, t)
}
}
