% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_air_humidity_2.R
\name{rh_calculation}
\alias{rh_calculation}
\title{Relative humidity (rh) calculation}
\usage{
rh_calculation(tmin, t)
}
\arguments{
\item{tmin}{A dataframe with minimum daily air temperature (°C)}

\item{t}{A dataframe with air temperature (°C) of the moment that you want to calculate the relative humidity.}
}
\value{
A data.frame object with the relative humidity calculated
}
\description{
Relative humidity is calculated in this function based on minimum air temperature of the day and the air temperature of the moment.
}
\examples{
\dontrun{
rh <- rh_calculation(tmin, t)
}
}
\author{
Roberto Filgueiras, Luan P. Venancio, Catariny C. Aleman and Fernando F. da Cunha
}
