% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation_parameters.R
\name{rn_calculation}
\alias{rn_calculation}
\title{Net radiation (rn)
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
rn_calculation(rns, rnl)
}
\arguments{
\item{rns}{The incomimg net shortwave radiation (MJ m-2 day-1).}

\item{rnl}{The outgoing net longwave radiation (MJ m-2 day-1).}
}
\description{
The net radiation (MJ m-2 day-1) is the difference between the incoming net shortwave radiation (rns) and the outgoing net longwave radiation (rnl).
}
\examples{
\dontrun{
rn <- rn_calculation(rns, rnl)
}
}
