% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_stations_info.R
\name{see_stations_info}
\alias{see_stations_info}
\title{Localization of the automatic weather station of INMET
\if{html}{\figure{logo_BrazilMet.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
see_stations_info()
}
\value{
A dataframe with informations of OMM code, latitude, longitude and altitude of all AWS stations available in INMET.
}
\description{
Function to see the localization of the automatic weather station of INMET.
}
\examples{
\dontrun{
see_stations_info()
}
}
\author{
Roberto Filgueiras
}
